Maybe makes something work with RSCs and React Query??
