+++ src/app/tricks/page.tsx
@@ -1,24 +1,17 @@
-'use client';
 import React from 'react';
 
-import { useQuery } from '@tanstack/react-query';
 import Tricktionary from '@/components/tricktionary/tricktionary';
+import Trick from '@/models/trick/trick';
 
 async function getTricks() {
-	return fetch('api/tricks').then((response) => response.json());
+	return fetch(`${process.env.API_URL}/tricks`).then((response) => {
+		return response.json() as unknown as Trick[];
+	});
 }
 
-export default function TricksList() {
-	const tricksQuery = useQuery({
-		queryKey: ['tricks'],
-		queryFn: getTricks,
-	});
-
-	if (tricksQuery.isLoading) return <h1>loading...</h1>;
-	if (tricksQuery.isError)
-		return <pre>{JSON.stringify(tricksQuery.error)}</pre>;
-
-	const tricks = tricksQuery.data;
+export default async function TricksList() {
+	const tricks = await getTricks();
+	console.log(tricks);
 
 	return <Tricktionary tricks={tricks} />;
 }
